cyclops equivalent output tests
Added tests showing that a initialized multi-hot model and a initialized standard model will return the same output if the multi-hot model is provided with an all-zero multi-hot encoding along with the input, or missing is provided along with the input, or nothing is provided along with the input. Also removed some variable names and moved things directly into the tests. Working on cyclops function step by step next.